migrate zip lib to archiver
