+++ lib/formats/gzip.js
@@ -1,38 +1,12 @@
-const { createWriteStream } = require('node:fs');
 const archiver = require('archiver');
+const { compress } = require('./archiver');
 
 module.exports = {
   label: 'Gzip',
   extension: 'tar.gz',
   output(filepath, data) {
-    return new Promise((resolve, reject) => {
-      const archive = archiver('tar', { gzip: true });
-      const output = createWriteStream(filepath);
+    const archive = archiver('tar', { gzip: true });
 
-      archive.on('warning', function(err) {
-        if (err.code === 'ENOENT') {
-          console.warn(err);
-        } else {
-          reject(err);
-        }
-      });
-      archive.on('error', reject);
-      archive.on('finish', resolve);
-
-      archive.pipe(output);
-
-      for (const filename in data) {
-        const content = data[filename];
-
-        if (content.endsWith('/')) {
-          archive.directory(content, filename);
-          continue;
-        }
-
-        archive.append(content, { name: filename });
-      }
-
-      archive.finalize();
-    });
+    return compress(filepath, data, archive);
   }
 };
