+++ lib/formats/zip.js
@@ -1,22 +1,11 @@
-const Zip = require('adm-zip');
+const archiver = require('archiver');
+const { compress } = require('./archiver');
 
 module.exports = {
   label: 'Zip',
   output(filepath, data) {
-    const zip = new Zip();
+    const archive = archiver('zip');
 
-    for (const filename in data) {
-      const content = data[filename];
-
-      try {
-        zip.addFile(filename, content);
-      } catch (error) {
-        self.apos.util.error('exportRecord error', error);
-      }
-    }
-
-    zip.writeZip(filepath);
-
-    return zip.toBuffer();
+    return compress(filepath, data, archive);
   }
 };
